server/requirements.txt
Django==2.0.1
django-extensions==1.9.8
djangorestframework==3.7.7
django-rest-swagger==2.1.2
django-model-utils==3.1.2
django-extra-fields==1.0.0
Pillow==5.1.0
psycopg2

pylint==1.7.5
pylint-django==0.7.2
django-debug-toolbar==1.8
ipython
pudb
